Přejmenování __init__.py na main.py, test_main.py
